requirements.txt
python-dotenv==1.0.1
openai==1.53.0
elevenlabs==1.5.0
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5